move utils to visual_gen
