+++ examples/visual_gen/visual_gen_ltx2.py
@@ -7,15 +7,14 @@
 import argparse
 import time
 
-from example_utils import (
+from tensorrt_llm import VisualGen, VisualGenArgs, VisualGenParams, logger
+from tensorrt_llm._torch.visual_gen.example_utils import (
     add_attention_backend_args,
     add_cache_dit_args,
     add_optimization_args,
     add_quant_args,
     build_cache_dit_config,
 )
-
-from tensorrt_llm import VisualGen, VisualGenArgs, VisualGenParams, logger
 from tensorrt_llm._torch.visual_gen.utils import linear_type_to_quant_config
 from tensorrt_llm.serve.media_storage import MediaStorage
